wildguard-backend/requirements.txt
# WildGuard AI — Backend Dependencies

# ── Core ────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9        # file uploads
httpx==0.27.0                  # fetch images from URLs
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Database ────────────────────────────────────────────
aiosqlite==0.20.0              # local SQLite (dev / offline)
# supabase==2.4.6              # uncomment for Supabase (production)

# ── AI / ML ─────────────────────────────────────────────
ultralytics==8.2.18            # YOLOv8
Pillow==10.3.0                 # image handling
transformers==4.40.2           # HuggingFace species classifier
torch==2.3.0                   # PyTorch (CPU build — swap for cu121 on GPU)
torchvision==0.18.0

# ── Alerts ──────────────────────────────────────────────
twilio==9.0.5                  # SMS alerts (optional — console fallback if not configured)

# ── Dev ─────────────────────────────────────────────────
python-dotenv==1.0.1
